refactor: altera namespace de job e fila
