+++ app/Infrastructure/Job/Queue/AsyncWithdrawConsumerJob.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Job;
+namespace App\Infrastructure\Job\Queue;
 
 use App\Application\Withdraw\Withdrawer;
 use App\Domain\Entity\Withdrawal;
